@@ -1,40 +1,91 @@
 {
-  "name": "mentorax-platform",
-  "version": "1.0.0",
-  "description": "Complete educational platform with admin capabilities",
+  "name": "brainex-platform",
+  "version": "2.0.0",
+  "description": "BraineX - Scholarships, Mentorship & Global Opportunities Platform",
   "main": "server.js",
+  "type": "module",
   "scripts": {
-    "start": "node server.js",
-    "dev": "nodemon server.js",
-    "install-deps": "npm install",
-    "setup": "npm install && echo '✅ Setup complete! Run: npm start'"
+    "dev": "concurrently \"npm run server:dev\" \"npm run client:dev\"",
+    "server:dev": "nodemon server.js",
+    "client:dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview",
+    "test": "jest --coverage",
+    "test:watch": "jest --watch",
+    "test:e2e": "cypress run",
+    "test:e2e:open": "cypress open",
+    "test:visual": "playwright test",
+    "test:visual:update": "playwright test --update-snapshots",
+    "lint": "eslint . --ext .js,.jsx",
+    "lint:fix": "eslint . --ext .js,.jsx --fix",
+    "format": "prettier --write \"**/*.{js,jsx,json,css,html}\"",
+    "prepare": "husky install",
+    "migrate": "node database/migrate.js",
+    "seed": "node database/seed.js",
+    "docker:build": "docker-compose build",
+    "docker:up": "docker-compose up -d",
+    "docker:down": "docker-compose down",
+    "i18n:extract": "node scripts/extract-translations.js",
+    "test:a11y": "node tests/accessibility.test.js",
+    "test:perf": "node tests/performance.test.js"
   },
   "dependencies": {
-    "bcryptjs": "^3.0.3",
-    "compression": "^1.8.1",
-    "cors": "^2.8.5",
-    "dotenv": "^16.4.5",
     "express": "^4.18.2",
-    "express-rate-limit": "^8.2.1",
-    "express-validator": "^7.3.1",
-    "helmet": "^8.1.0",
-    "jsonwebtoken": "^9.0.3",
-    "mongoose": "^9.1.1",
-    "node-fetch": "^3.3.2",
-    "nodemailer": "^7.0.12",
-    "nodemon": "^3.0.1",
-    "validator": "^13.15.26"
+    "express-rate-limit": "^7.1.5",
+    "bcrypt": "^5.1.1",
+    "jsonwebtoken": "^9.0.2",
+    "joi": "^17.11.0",
+    "helmet": "^7.1.0",
+    "cors": "^2.8.5",
+    "dotenv": "^16.3.1",
+    "mysql2": "^3.6.5",
+    "multer": "^1.4.5-lts.1",
+    "nodemailer": "^6.9.7",
+    "socket.io": "^4.6.0",
+    "compression": "^1.7.4",
+    "express-validator": "^7.0.1",
+    "cookie-parser": "^1.4.6",
+    "csurf": "^1.11.0",
+    "sharp": "^0.33.1",
+    "bull": "^4.12.0",
+    "winston": "^3.11.0",
+    "morgan": "^1.10.0"
+  },
+  "devDependencies": {
+    "vite": "^5.0.10",
+    "nodemon": "^3.0.2",
+    "concurrently": "^8.2.2",
+    "eslint": "^8.56.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-node": "^11.1.0",
+    "prettier": "^3.1.1",
+    "husky": "^8.0.3",
+    "lint-staged": "^15.2.0",
+    "jest": "^29.7.0",
+    "supertest": "^6.3.3",
+    "cypress": "^13.6.2",
+    "@testing-library/jest-dom": "^6.1.5",
+    "jest-environment-jsdom": "^29.7.0"
+  },
+  "engines": {
+    "node": ">=18.0.0",
+    "npm": ">=9.0.0"
+  },
+  "lint-staged": {
+    "*.js": [
+      "eslint --fix",
+      "prettier --write"
+    ],
+    "*.{json,css,html}": [
+      "prettier --write"
+    ]
   },
   "keywords": [
-    "education",
     "scholarships",
-    "mentors",
-    "admin-panel",
-    "student-platform"
+    "mentorship",
+    "education",
+    "opportunities"
   ],
-  "author": "MentoraX Team",
-  "license": "MIT",
-  "engines": {
-    "node": ">=14.0.0"
-  }
+  "author": "BraineX Team",
+  "license": "MIT"
 }
