@@ -1,42 +1,43 @@
-// Database configuration
-const mongoose = require('mongoose');
+import mysql from 'mysql2/promise';
+import dotenv from 'dotenv';
 
-const connectDB = async () => {
+dotenv.config();
+
+// Create connection pool for better performance
+export const pool = mysql.createPool({
+    host: process.env.DB_HOST || 'localhost',
+    port: process.env.DB_PORT || 3306,
+    user: process.env.DB_USER || 'root',
+    password: process.env.DB_PASSWORD || '',
+    database: process.env.DB_NAME || 'brainex_db',
+    waitForConnections: true,
+    connectionLimit: parseInt(process.env.DB_CONNECTION_LIMIT) || 10,
+    queueLimit: 0,
+    enableKeepAlive: true,
+    keepAliveInitialDelay: 0,
+});
+
+// Test database connection
+export const testConnection = async () => {
     try {
-        const mongoURI = process.env.MONGODB_URI || 'mongodb://localhost:27017/brainex';
-
-        const conn = await mongoose.connect(mongoURI, {
-            // Mongoose 8 defaults are good, but we'll be explicit
-            maxPoolSize: 10,
-            serverSelectionTimeoutMS: 5000,
-            socketTimeoutMS: 45000,
-        });
-
-        console.log(`📦 MongoDB Connected: ${conn.connection.host}`);
-
-        // Handle connection events
-        mongoose.connection.on('error', (err) => {
-            console.error('MongoDB connection error:', err);
-        });
-
-        mongoose.connection.on('disconnected', () => {
-            console.log('MongoDB disconnected');
-        });
-
-        // Graceful shutdown
-        process.on('SIGINT', async () => {
-            await mongoose.connection.close();
-            console.log('MongoDB connection closed due to app termination');
-            process.exit(0);
-        });
-
-        return conn;
+        const connection = await pool.getConnection();
+        console.log('✅ Database connected successfully');
+        connection.release();
+        return true;
     } catch (error) {
-        console.warn('⚠️  MongoDB connection failed:', error.message);
-        console.log('🔧 Running in DEMO MODE without database');
-        console.log('   (User registration/login will work but data won\'t persist)');
-        return null;
+        console.error('❌ Database connection failed:', error.message);
+        return false;
     }
 };
 
-module.exports = connectDB;
+// Graceful shutdown
+export const closePool = async () => {
+    try {
+        await pool.end();
+        console.log('Database pool closed');
+    } catch (error) {
+        console.error('Error closing database pool:', error);
+    }
+};
+
+export default pool;
