@@ -1,33 +1,39 @@
-const Application = require('../models/Application');
-const memoryStore = require('../utils/memoryStore');
-const mongoose = require('mongoose');
+import Application from '../models/Application.js';
+import { pool } from '../config/database.js';
 
-const isMongoConnected = () => mongoose.connection.readyState === 1;
-
-exports.createApplication = async (req, res) => {
+export const createApplication = async (req, res) => {
     try {
         const { scholarshipId, mentorId, type, data } = req.body;
         const userId = req.user.id; // From auth middleware
 
-        if (isMongoConnected()) {
-            // Check for existing
-            const query = { userId, type };
-            if (scholarshipId) query.scholarshipId = scholarshipId;
-            if (mentorId) query.mentorId = mentorId;
+        // Check for existing application
+        const [existing] = await pool.query(
+            'SELECT id FROM applications WHERE user_id = ? AND type = ?',
+            [userId, type]
+        );
 
-            const existing = await Application.findOne(query);
-            if (existing) {
-                return res.status(400).json({ success: false, error: 'You have already applied.' });
-            }
+        if (existing.length > 0) {
+            return res.status(400).json({ success: false, error: 'You have already applied.' });
+        }
 
-            const application = await Application.create({
+        const [result] = await pool.query(
+            'INSERT INTO applications (user_id, scholarship_id, mentor_id, type, data, status) VALUES (?, ?, ?, ?, ?, ?)',
+            [userId, scholarshipId || null, mentorId || null, type, JSON.stringify(data), 'pending']
+        );
+
+        return res.status(201).json({ 
+            success: true, 
+            data: {
+                id: result.insertId,
                 userId,
                 scholarshipId,
                 mentorId,
                 type,
-                data
-            });
-            return res.status(201).json({ success: true, data: application });
+                data,
+                status: 'pending',
+                submittedAt: new Date()
+            }
+        });
         } else {
             // Memory Store
             const existing = memoryStore.applications.find(a =>
@@ -60,35 +66,40 @@ exports.createApplication = async (req, res) => {
     }
 };
 
-exports.getUserApplications = async (req, res) => {
+export const getUserApplications = async (req, res) => {
     try {
         const userId = req.user.id;
-        let apps = [];
 
-        if (isMongoConnected()) {
-            apps = await Application.find({ userId })
-                .populate('scholarshipId')
-                .populate('mentorId')
-                .sort({ submittedAt: -1 });
-        } else {
-            apps = memoryStore.applications.filter(a => a.userId === userId || a.userId === req.user._id);
-            // Manually populate (simulate)
-            apps = apps.map(app => {
-                let populated = { ...app };
-                if (app.scholarshipId) {
-                    populated.scholarshipId = memoryStore.scholarships.find(s => s._id === app.scholarshipId || s.id === app.scholarshipId);
-                }
-                if (app.mentorId) {
-                    populated.mentorId = memoryStore.mentors.find(m => m._id === app.mentorId || m.id === app.mentorId);
-                }
-                return populated;
-            });
-            apps.sort((a, b) => new Date(b.submittedAt) - new Date(a.submittedAt));
-        }
+        const [apps] = await pool.query(
+            `SELECT a.id, a.user_id, a.scholarship_id, a.mentor_id, a.type, a.data, 
+                    a.status, a.submitted_at, s.name as scholarship_name
+             FROM applications a
+             LEFT JOIN scholarships s ON a.scholarship_id = s.id
+             WHERE a.user_id = ?
+             ORDER BY a.submitted_at DESC`,
+            [userId]
+        );
 
-        res.json({ success: true, count: apps.length, data: apps });
+        const formattedApps = apps.map(app => ({
+            id: app.id,
+            userId: app.user_id,
+            scholarshipId: app.scholarship_id,
+            mentorId: app.mentor_id,
+            type: app.type,
+            data: typeof app.data === 'string' ? JSON.parse(app.data) : app.data,
+            status: app.status,
+            submittedAt: app.submitted_at,
+            scholarshipData: app.scholarship_name ? { name: app.scholarship_name } : null
+        }));
+
+        res.json({ success: true, count: formattedApps.length, data: formattedApps });
     } catch (error) {
         console.error('Get Applications Error:', error);
         res.status(500).json({ success: false, error: 'Server Error' });
     }
+};
+
+export default {
+    createApplication,
+    getUserApplications
 };
