@@ -1,58 +1,17 @@
-// Authentication Routes - Production-ready
-const express = require('express');
+import express from 'express';
+import { register, login, logout, getCurrentUser, refreshAccessToken } from '../controllers/authController.js';
+import { authenticate } from '../middleware/auth.js';
+import { validate, registerSchema, loginSchema } from '../utils/validation.js';
+
 const router = express.Router();
-const { body } = require('express-validator');
-const authController = require('../controllers/authController');
-const { authenticate } = require('../middleware/auth');
-
-// Validation rules
-const registerValidation = [
-    body('firstName')
-        .trim()
-        .notEmpty().withMessage('First name is required')
-        .isLength({ max: 50 }).withMessage('First name cannot exceed 50 characters'),
-    body('lastName')
-        .trim()
-        .notEmpty().withMessage('Last name is required')
-        .isLength({ max: 50 }).withMessage('Last name cannot exceed 50 characters'),
-    body('email')
-        .trim()
-        .notEmpty().withMessage('Email is required')
-        .isEmail().withMessage('Please provide a valid email')
-        .normalizeEmail(),
-    body('password')
-        .notEmpty().withMessage('Password is required')
-        .isLength({ min: 8 }).withMessage('Password must be at least 8 characters')
-        .matches(/^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)/).withMessage('Password must contain uppercase, lowercase, and number'),
-    body('field')
-        .optional()
-        .isIn(['ai', 'biotech', 'climate', 'engineering', 'entrepreneurship', 'social', 'media', 'economics', 'other'])
-        .withMessage('Invalid field selection')
-];
-
-const loginValidation = [
-    body('email')
-        .trim()
-        .notEmpty().withMessage('Email is required')
-        .isEmail().withMessage('Please provide a valid email'),
-    body('password')
-        .notEmpty().withMessage('Password is required')
-];
 
 // Public routes
-router.post('/register', registerValidation, authController.register);
-router.post('/login', loginValidation, authController.login);
-router.post('/refresh-token', authController.refreshToken);
-router.post('/forgot-password', authController.forgotPassword);
-router.get('/verify-email/:token', authController.verifyEmail);
+router.post('/register', validate(registerSchema), register);
+router.post('/login', validate(loginSchema), login);
+router.post('/logout', logout);
+router.post('/refresh', refreshAccessToken);
 
 // Protected routes
-router.get('/me', authenticate, authController.getMe);
-router.post('/logout', authenticate, authController.logout);
-router.put('/profile', authenticate, authController.updateProfile);
-router.put('/change-password', authenticate, authController.changePassword);
+router.get('/me', authenticate, getCurrentUser);
 
-// Legacy endpoint for backward compatibility
-router.get('/current-user', authenticate, authController.getMe);
-
-module.exports = router;
+export default router;
