@@ -1,11 +1,23 @@
-const express = require('express');
+import express from 'express';
+import {
+    getScholarships,
+    getScholarshipById,
+    createScholarship,
+    updateScholarship,
+    deleteScholarship,
+} from '../controllers/scholarshipController.js';
+import { authenticate, authorize } from '../middleware/auth.js';
+import { validate, scholarshipSchema, validateId } from '../utils/validation.js';
+
 const router = express.Router();
-const scholarshipController = require('../controllers/scholarshipController');
 
-// Get all scholarships (public route)
-router.get('/', scholarshipController.getScholarships);
+// Public routes
+router.get('/', getScholarships);
+router.get('/:id', validateId, getScholarshipById);
 
-// Get scholarship by ID
-router.get('/:id', scholarshipController.getScholarshipById);
+// Admin routes
+router.post('/', authenticate, authorize('admin'), validate(scholarshipSchema), createScholarship);
+router.put('/:id', authenticate, authorize('admin'), validateId, updateScholarship);
+router.delete('/:id', authenticate, authorize('admin'), validateId, deleteScholarship);
 
-module.exports = router;
+export default router;
