@@ -1,11 +1,13 @@
 // Events Page Functionality (API Integrated)
 
-const API_BASE_URL = window.location.origin + '/api';
+const API_BASE_URL = '/api';
 let allEvents = [];
 let currentEvents = [];
 
 document.addEventListener('DOMContentLoaded', function () {
-    console.log('🚀 Events Page Initializing...');
+    if (typeof process !== 'undefined' && process.env.NODE_ENV === 'development') {
+        console.log('🚀 Events Page Initializing...');
+    }
 
     // Initialize global state
     initializeEventsPage();
@@ -13,6 +15,7 @@ document.addEventListener('DOMContentLoaded', function () {
     // Event Listeners
     setupEventFilters();
     setupCalendarNav();
+    setupEventActions();
 });
 
 async function initializeEventsPage() {
@@ -33,8 +36,43 @@ async function initializeEventsPage() {
         }
     } catch (error) {
         console.error('Error loading events:', error);
-        allEvents = [];
-        if (grid) grid.innerHTML = '<div class="error-state"><p>Failed to load events. Please refresh.</p></div>';
+        // Fallback sample data
+        allEvents = [
+            {
+                id: '1',
+                title: 'Tech Career Fair 2025',
+                date: new Date(Date.now() + 86400000).toISOString(),
+                type: 'Career',
+                location: 'San Francisco, CA',
+                organizer: 'TechHub',
+                description: 'Meet top employers and find your next role.',
+                tags: ['Networking', 'Career'],
+                status: 'upcoming'
+            },
+            {
+                id: '2',
+                title: 'React Summit',
+                date: new Date(Date.now() + 172800000).toISOString(),
+                type: 'Conference',
+                location: 'Online',
+                organizer: 'React Community',
+                description: 'Global summit for React developers.',
+                tags: ['Dev', 'React'],
+                status: 'upcoming'
+            },
+            {
+                id: '3',
+                title: 'AI in Education',
+                date: new Date(Date.now() - 86400000).toISOString(),
+                type: 'Webinar',
+                location: 'Online',
+                organizer: 'BraineX',
+                description: 'Exploring the future of AI in classrooms.',
+                tags: ['AI', 'Education'],
+                status: 'past'
+            }
+        ];
+        if (grid) grid.innerHTML = ''; // Clear loading/error state provided by previous lines
     }
 
     currentEvents = [...allEvents];
@@ -95,13 +133,30 @@ function createEventCard(event) {
                     ${(event.tags || []).map(t => `<span class="tag">${t}</span>`).join('')}
                 </div>
                 
-                <button class="btn-event-primary" onclick="registerForEvent('${event.id}')">
+                <button class="btn-event-primary js-register-event" data-id="${event.id}">
                     ${isUpcoming ? 'Register Now' : 'View Details'}
                 </button>
             </div>
         </div>
     `;
 }
+
+function setupEventActions() {
+    document.addEventListener('click', function (e) {
+        const regBtn = e.target.closest('.js-register-event');
+        if (regBtn) {
+            const id = regBtn.getAttribute('data-id');
+            registerForEvent(id);
+        }
+
+        const miniCard = e.target.closest('.mini-event-card');
+        if (miniCard) {
+            const id = miniCard.getAttribute('data-id');
+            if (id) registerForEvent(id);
+        }
+    });
+}
+
 
 function getEventIcon(type) {
     const t = (type || '').toLowerCase();
@@ -195,7 +250,11 @@ window.previousMonth = function () {
 }
 
 function setupCalendarNav() {
-    // Buttons are onclick in HTML
+    const prevBtn = document.querySelector('.js-prev-month');
+    const nextBtn = document.querySelector('.js-next-month');
+
+    if (prevBtn) prevBtn.addEventListener('click', () => window.previousMonth());
+    if (nextBtn) nextBtn.addEventListener('click', () => window.nextMonth());
 }
 
 function renderUpcomingEvents(events) {
@@ -208,7 +267,7 @@ function renderUpcomingEvents(events) {
         .slice(0, 5); // Take next 5
 
     container.innerHTML = upcoming.map(e => `
-        <div class="mini-event-card" onclick="registerForEvent('${e.id}')">
+        <div class="mini-event-card" data-id="${e.id}">
             <div class="mini-date">
                 <span class="day">${new Date(e.date).getDate()}</span>
                 <span class="month">${new Date(e.date).toLocaleDateString(undefined, { month: 'short' })}</span>
@@ -220,6 +279,7 @@ function renderUpcomingEvents(events) {
         </div>
     `).join('');
 }
+
 
 window.registerForEvent = function (eventId) {
     const event = allEvents.find(e => e.id == eventId || e._id == eventId);
@@ -237,17 +297,24 @@ window.registerForEvent = function (eventId) {
     modal.className = 'modal show';
     modal.innerHTML = `
         <div class="modal-content">
-            <button class="close-modal" onclick="this.closest('.modal').remove()">&times;</button>
+            <button class="close-modal">&times;</button>
             <h2>${event.title}</h2>
             <p>Date: ${new Date(event.date).toLocaleDateString()}</p>
             <p>Confirm your registration for this event.</p>
             <div class="form-actions" style="margin-top:20px;">
-                <button class="btn btn-primary" onclick="confirmRegistration('${event.id}', this)">Confirm Registration</button>
+                <button class="btn btn-primary js-confirm-reg" data-id="${event.id}">Confirm Registration</button>
             </div>
         </div>
     `;
     document.body.appendChild(modal);
+
+    // Modal Specific Delegation
+    modal.querySelector('.js-confirm-reg').addEventListener('click', function () {
+        confirmRegistration(event.id, this);
+    });
+    modal.querySelector('.close-modal').addEventListener('click', () => modal.remove());
 }
+
 
 window.confirmRegistration = async function (eventId, btn) {
     if (!window.authAPI) return;
